Fixed application/octet-stream to guess mimetype
When being downloaded, some files were being classified as
application/octet-stream in the response even though they are
a valid mimetype. This adds in guessing the mimetype before
sending off the intent. It was chosen to put this here
instead of when it is downloaded to make it backwards compatible
with files that already exist in people's downloads.

Change-Id: Id5fc9a3cca6851677b066854fcdccb5ae4fade7c
Issue-Id: CYNGNOS-1153

Patchset: 3
http://review.cyanogenmod.org/#/c/110515/